Add functionality for CVE Lookup Utility License ========================== (SPDX: MIT)
(c) 2024 Gerald Hilts

Permission is hereby granted, free of charge, to any person obtaining a copy of
this software and associated documentation files (the “Software”), to deal in
the Software without restriction, including without limitation the rights to
use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
the Software, and to permit persons to whom the Software is furnished to do so,
subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

(https://github.com/gwhilts/cve_lookup/blob/main/LICENSE)

NIST NVD
========

The CVE Lookup Utility uses application programming interfaces (APIs) to
retreive information from the National Vulnerability Database (NVD), a U.S.
government repository of standards based vulnerability management data.

This data is made available by the National Institute of Standards and
Technology (NIST) as a "as a public service" with the following legal
disclaimer:

  The National Vulnerability Database (NVD) is a repository of standards based
  vulnerability data. The Database is maintained by the National Institute of
  Standards and Technology (NIST), an agency of the Federal Government, and is
  being provided as a public service. Much of the data in NVD records is derived
  from publicly available data sources, including product information and
  manufacturer/developer information. NIST does not evaluate, review, or test
  software or code contained within the NVD. The NVD is expressly provided
  “AS IS.” NIST MAKES NO WARRANTY OF ANY KIND, EXPRESS, IMPLIED OR STATUTORY,
  INCLUDING, WITHOUT LIMITATION, THE IMPLIED WARRANTY OF MERCHANTABILITY, FITNESS
  FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND DATA ACCURACY. NIST does not
  warrant or make any representations regarding the use of the NVD, its contents,
  or the results obtained therefrom, including but not limited to the correctness,
  accuracy, reliability, or usefulness of the NVD. You are solely responsible for
  determining the appropriateness of your use of the NVD and its contents and you
  assume all risks associated with its use.

(see https://nvd.nist.gov/general/legal-disclaimer)

The CVE Lookup Utility is not endorsed or certified by the NVD.

Ruby Gems
=========

Although not present in the code itself, various Ruby Gem libraries
that CVE Lookup Utility rely upon may be installed along with it
if they are not already present in the user's environment.

These libraries are released under various licenses. See

  https://github.com/gwhilts/cve_lookup/blob/main/THIRD_PARTY.md

for more details. command